项目工作结束

+++ demo2/src/main/java/com/example/demo/entity/Count.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class Count {
     public int usercount;
     public int clockincount;
-    public int clockintime;
+    public float clockintime;
 
 }
